+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [1.13.0] - 2023-11-12
+
 ### Added
 
 - Add new `init` command to ensure configurations are available without running the farmer.
@@ -163,7 +165,8 @@
 
 - Initial release
 
-[unreleased]: https://github.com/foxypool/foxy-farmer/compare/1.12.1...HEAD
+[unreleased]: https://github.com/foxypool/foxy-farmer/compare/1.13.0...HEAD
+[1.13.0]: https://github.com/foxypool/foxy-farmer/compare/1.12.1...1.13.0
 [1.12.1]: https://github.com/foxypool/foxy-farmer/compare/1.12.0...1.12.1
 [1.12.0]: https://github.com/foxypool/foxy-farmer/compare/1.11.0...1.12.0
 [1.11.0]: https://github.com/foxypool/foxy-farmer/compare/1.10.2...1.11.0